+++ playground/Sample/Program.cs
@@ -1,6 +1,0 @@
-﻿/*
- * To view how a simple command works, check Commands\Hello.cs.
- * To view how the command-hierarchy works, check Group\*
- */
-
-App.Run(new(args));